don't track virtual environment or (for now) PyCharm project files
The virtual environment will never be added.

PyCharm project files may be added later if they turn out to be useful on multiple machines or to multiple users.

In particular, PyCharm has a spellchecker that turns up mostly false positives, but it has found one true misspelling. Tracking the PyCharm project would avoid manually rebuilds of the dictionary.

(The misspelling it did find will be corrected in a subsequent commit.)